# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ArticleResponse": {
        "properties": {
          "author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Name"
          },
          "body_html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Html"
          },
          "body_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Text"
          },
          "content_id": {
            "title": "Content Id",
            "type": "string"
          },
          "external_url": {
            "title": "External Url",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "translated_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Title"
          },
          "translation_locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Locale"
          },
          "translation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Status"
          }
        },
        "required": [
          "content_id",
          "title",
          "translated_title",
          "translation_locale",
          "translation_status",
          "body_html",
          "body_text",
          "external_url",
          "author_name",
          "published_at"
        ],
        "title": "ArticleResponse",
        "type": "object"
      },
...
      "FeedItemResponse": {
        "properties": {
          "author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Name"
          },
          "content_id": {
            "title": "Content Id",
            "type": "string"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "external_url": {
            "title": "External Url",
            "type": "string"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "is_saved": {
            "title": "Is Saved",
            "type": "boolean"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "source_avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Avatar Url"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "translated_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Title"
          },
          "translation_locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Locale"
          },
          "translation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Status"
          },
          "x_preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XPreview"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "content_id",
          "source_id",
          "source_name",
          "source_avatar_url",
          "source_kind",
          "title",
          "translated_title",
          "translation_locale",
          "translation_status",
          "excerpt",
          "external_url",
          "author_name",
          "published_at",
          "fetched_at",
          "thumbnail_url",
          "is_saved"
        ],
        "title": "FeedItemResponse",
        "type": "object"
      },
      "FeedPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "FeedPageResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "XAuthor": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "XAuthor",
        "type": "object"
      },
      "XPreview": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/XAuthor"
          },
          "completeness": {
            "enum": [
              "parsed",
              "partial",
              "legacy"
            ],
            "title": "Completeness",
            "type": "string"
          },
          "external_url": {
            "title": "External Url",
            "type": "string"
          },
          "is_repost": {
            "default": false,
            "title": "Is Repost",
            "type": "boolean"
          },
          "quote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XReferencePreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "reply_to": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XReplyPreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "repost": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XReferencePreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "thread": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XThreadPreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "tweet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tweet Id"
          }
        },
        "required": [
          "tweet_id",
          "text",
          "author",
          "external_url",
          "completeness"
        ],
        "title": "XPreview",
        "type": "object"
      },
      "XReferencePreview": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/XAuthor"
          },
          "availability": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Availability",
            "type": "string"
          },
          "external_url": {
            "title": "External Url",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "tweet_id": {
            "title": "Tweet Id",
            "type": "string"
          }
        },
        "required": [
          "tweet_id",
          "author",
          "external_url",
          "availability"
        ],
        "title": "XReferencePreview",
        "type": "object"
      },
      "XReplyPreview": {
        "properties": {
          "author_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Id"
          },
          "handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "tweet_id": {
            "title": "Tweet Id",
            "type": "string"
          }
        },
        "required": [
          "tweet_id"
        ],
        "title": "XReplyPreview",
        "type": "object"
      },
      "XThreadPreview": {
        "properties": {
          "loaded_count": {
            "title": "Loaded Count",
            "type": "integer"
          },
          "root_tweet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Tweet Id"
          },
          "root_url": {
            "title": "Root Url",
            "type": "string"
          }
        },
        "required": [
          "loaded_count",
          "root_tweet_id",
          "root_url"
        ],
        "title": "XThreadPreview",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "ReaderServerToken": {
        "description": "Deployment connection token configured by the server administrator.",
        "in": "header",
        "name": "X-Reader-Server-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend for shared-source information aggregation.",
    "title": "Reader API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/feed": {
      "get": {
        "description": "Return content only from sources to which the caller is subscribed.",
        "operationId": "list_feed_v1_feed_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "folder_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder Name"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedItemResponse"
                  },
                  "title": "Response List Feed V1 Feed Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Feed",
        "tags": [
          "feed"
        ]
      }
    },
    "/v1/feed/page": {
      "get": {
        "description": "Return a stable, deduplicated page of the caller's feed.",
        "operationId": "list_feed_page_v1_feed_page_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "folder_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder Name"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Feed Page",
        "tags": [
          "feed"
        ]
      }
    },
    "/v1/feed/{content_id}": {
      "get": {
        "description": "Return an article from an active subscription or the caller's saved list.",
        "operationId": "get_article_v1_feed__content_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Article",
        "tags": [
          "feed"
        ]
      }
    },
...
  }
}